requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.0.0
